+++ src/components/route-points.js
@@ -1,7 +1,6 @@
-import {formatTime} from "../utils";
-import {typeRoutePointMap, MAX_SHOWED_OFFERS_COUNT} from "./constants";
-import {getTimeDifference} from "./date-generation";
+import {typeRoutePointMap, MAX_SHOWED_OFFERS_COUNT} from "../constants";
 import AbstractComponent from "./abstract-component";
+import {getTimeDifference, formatTime} from "../utils/date-utils";
 
 const createOffersMarkup = (offers) => offers.slice(0, MAX_SHOWED_OFFERS_COUNT).map((offer) => {
   return (
