--- src/utils.js
@@ -1,77 +1,0 @@
-import {RenderPosition} from "./components/constants";
-
-export const getRandomInteger = (max, min = 0) => {
-  let rand = min + Math.random() * (max + 1 - min);
-  return Math.floor(rand);
-};
-
-export const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomInteger(array.length - 1);
-
-  return array[randomIndex];
-};
-
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-export const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours());
-  const minutes = castTimeFormat(date.getMinutes());
-
-  return `${hours}:${minutes}`;
-};
-
-export const getRandomDate = () => {
-  const targetDate = new Date();
-  const sign = Math.random() > 0.5 ? 1 : -1;
-  const diffValue = sign * getRandomInteger(10, 1);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-  targetDate.setHours(targetDate.getHours() + diffValue);
-  targetDate.setMinutes(targetDate.getMinutes() + diffValue);
-
-  return targetDate;
-};
-
-export const formatDate = (date) => {
-  const dd = (`0` + date.getDate()).slice(-2);
-  const mm = (`0` + (date.getMonth() + 1)).slice(-2);
-  const yy = date.getFullYear() % 100;
-
-  return `${dd}/${mm}/${yy}`;
-};
-
-export const datesArray = [];
-
-export const getAllDates = (defaultArray) => {
-  defaultArray.forEach((element) => {
-    return datesArray.push(element.dateFrom.toDateString());
-  });
-};
-
-export const getUniqueDates = (array) => {
-  return Array.from(new Set(array));
-};
-
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place = RenderPosition.BEFOREEND) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
