+++ src/utils/common.js
@@ -1,0 +1,33 @@
+import {RenderPosition} from "../constants";
+
+export const getRandomInteger = (max, min = 0) => {
+  let rand = min + Math.random() * (max + 1 - min);
+  return Math.floor(rand);
+};
+
+export const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomInteger(array.length - 1);
+
+  return array[randomIndex];
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+export const render = (container, element, place = RenderPosition.BEFOREEND) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(element);
+      break;
+    case RenderPosition.AFTEREND:
+      container.after(element);
+      break;
+  }
+};
