+++ src/utils/date-utils.js
@@ -1,5 +1,5 @@
-import {getRandomInteger, getRandomDate} from "../utils";
-import {Time} from "./constants";
+import {Time} from "../constants";
+import {getRandomInteger} from "./common";
 
 const MAX_COUNT_DAYS_DIFFERENCE = 1;
 
@@ -39,4 +39,47 @@ export const getTimeDifference = (start, end) => {
     return difference;
   }
   return difference;
+};
+
+const castTimeFormat = (value) => {
+  return value < 10 ? `0${value}` : String(value);
+};
+
+export const formatTime = (date) => {
+  const hours = castTimeFormat(date.getHours());
+  const minutes = castTimeFormat(date.getMinutes());
+
+  return `${hours}:${minutes}`;
+};
+
+export const getRandomDate = () => {
+  const targetDate = new Date();
+  const sign = Math.random() > 0.5 ? 1 : -1;
+  const diffValue = sign * getRandomInteger(10, 1);
+
+  targetDate.setDate(targetDate.getDate() + diffValue);
+  targetDate.setHours(targetDate.getHours() + diffValue);
+  targetDate.setMinutes(targetDate.getMinutes() + diffValue);
+
+  return targetDate;
+};
+
+export const formatDate = (date) => {
+  const dd = (`0` + date.getDate()).slice(-2);
+  const mm = (`0` + (date.getMonth() + 1)).slice(-2);
+  const yy = date.getFullYear() % 100;
+
+  return `${dd}/${mm}/${yy}`;
+};
+
+export const datesArray = [];
+
+export const getAllDates = (defaultArray) => {
+  defaultArray.forEach((element) => {
+    return datesArray.push(element.dateFrom.toDateString());
+  });
+};
+
+export const getUniqueDates = (array) => {
+  return Array.from(new Set(array));
 };
